Move progress bar helper into runtime-options boundary
Relocate src/cli/src/shared/progress-bar.js to src/cli/src/runtime-options/progress-bar.js so the CLI-only progress bar configuration lives with the other runtime option modules instead of the shared bundle. Update imports, exports, docs, and tests to follow the new location.
